pyproject.toml
[tool.poetry]
name = "toaster v2"
version = "0.1.0"
description = ""
authors = ["source64 <65923236+source64@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
motor = "^3.1.1"
aiohttp = "^3.8.3"
discord-py = "^2.1.0"
discord-ext-menus = {git = "https://github.com/Rapptz/discord-ext-menus"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
